# src/main/resources/mcShares.postman_collection-openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "",
      "variables": {}
    }
  ],
  "info": {
    "version": "589abd12-d08a-4c5c-afbe-e2b5b5eed02b",
    "title": "mcShares",
    "description": "",
    "termsOfService": "",
    "contact": {},
    "license": {
      "name": ""
    }
  },
  "paths": {
...
    "/mcshares/retrieveRecord": {
      "get": {
        "summary": "http://localhost:8080/mcshares/retrieveRecord",
        "operationId": "http://localhost:8080/mcshares/retrieveRecord",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJ0YXNod2VlbiIsImV4cCI6MTY0MDk3MjI5OCwiaWF0IjoxNjQwOTU0Mjk4fQ.YtG-fmcU7fFPHGRMpck0E8aQLRpmNz2-nG8a07ZoSXTkB_9Eycjt8y5k6HBKvr6wpXA9S1jFctWFOp7ale_N0w"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API to retrieve the specific customer information",
            "headers": {}
          }
        }
      }
    },
...
    "/mcshares/downloadDataAsCSV": {
      "get": {
        "summary": "http://localhost:8080/mcshares/downloadDataAsCSV",
        "operationId": "http://localhost:8080/mcshares/downloadDataAsCSV",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJ0YXNod2VlbiIsImV4cCI6MTY0MDk3MjI5OCwiaWF0IjoxNjQwOTU0Mjk4fQ.YtG-fmcU7fFPHGRMpck0E8aQLRpmNz2-nG8a07ZoSXTkB_9Eycjt8y5k6HBKvr6wpXA9S1jFctWFOp7ale_N0w"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API to generate and download customers records in a CSV file",
            "headers": {}
          }
        }
      }
    },
...
  },
...
  "security": [],
...
}